requirements.txt
# Core Framework
streamlit>=1.28.0

# AI & LLM
cerebras-cloud-sdk>=1.0.0

# Web Scraping
trafilatura>=1.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Document Generation
python-docx>=0.8.11

# Data Processing
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.2

python-dotenv>=1.0.0